+++ MRP_CharacterSettings.lua
@@ -2,9 +2,13 @@
 -- local _, MRP = ...
 
 ---@class CharacterSettings
+---@field autoSkip boolean Whether to automatically skip steps in the route.
+---@field autoAdvance boolean Whether to automatically advance to the next step after completing the current one.
 ---@field currentStep number The current step in the route being followed.
 ---@field filter FilterSettings The filter settings for the character.
 MRP_CharacterSettings = {
+    autoSkip = true,
+    autoAdvance = true,
     currentStep = 1,
     filter = {
         expansions = {