+++ MRP_Settings.lua
@@ -2,12 +2,11 @@
 -- local _, MRP = ...
 
 ---@class Settings
----@field autoSkip boolean Whether to automatically skip steps in the route.
----@field autoAdvance boolean Whether to automatically advance to the next step after completing the current one.
 ---@field useTomTom boolean Whether to use TomTom for waypoint navigation.
+---@field showDifficultyWarning boolean Whether to show a warning for incorrect difficulty settings.
+---@field ignoreLFRDifficulty boolean Whether to ignore LFR difficulty on steps.
 MRP_Settings = {
-    autoSkip = true,
-    autoAdvance = true,
     useTomTom = true,
     showDifficultyWarning = true,
+    ignoreLFRDifficulty = false
 }